Bug fixes regarding Camera movement.
I figured out the right way to enable/disable components, so I changed
that around and got rid of the 'enable' bool.  Also, we should have
been changing Character Motor, not FPS Input, and there were two Mouse
Look scripts, one on the main camera, and one on the player.  I assume
that's for a reason, so I just disable both now during a cutscene